requirements.txt
fitz
numpy
fastapi
uvicorn
python-multipart
pydantic
openai
sentence-transformers
faiss-cpu
PyPDF2
python-docx
streamlit
